# src/config/docs/swagger.yml
openapi: 3.0.3
info:
  version: 1.0.0
  termsOfService: 'http://swagger.io/terms/'
  title: Pay-Slips-API
  license:
    name: ISC
  contact:
    name: Alejandro Valverde
    email: avalverde@midasconsultores.com.ar
paths:
  '/register':
    post:
      tags:
        - Users
      summary: Register a new user
      description: This endpoint creates a new user
      operationId: RegisterUser
      requestBody:
        $ref: '#/components/requestBodies/CreateUser'
      responses:
        '200':
          $ref: '#/components/responses/GetUserResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
servers:
  - url: '{protocol}://{domain}{port}/{baseUrl}/{apiVersion}'
    description: Default
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      domain:
        default: localhost
      port:
        default: ':8080'
      baseUrl:
        default: 'api'
      apiVersion:
        enum:
          - 'v1'
        default: 'v1'
components:
  requestBodies:
    CreateUser:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserBody'
            required: true
...
  responses:
...
    GetUserResponse:
      description: Successful operation
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/StandardResponse'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Users'
                  message:
                    type: string
                    example: null
                    nullable: true
...
    BadRequestResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  status:
                    example: 400
                  message:
                    example: The date is invalid
    UnauthorizedResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  status:
                    example: 401
                  message:
                    example: Invalid username or password
    ForbiddenResponse:
      description: You do not have permission to access this resource
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  status:
                    example: 403
                  message:
                    example: You do not have permission to perform this action.
    InternalErrorResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  status:
                    example: 500
                  message:
                    example: Internal server error
...
  schemas:
    UserBody:
      type: object
      properties:
        cuil:
          type: string
          example: 23-12345678-1
        name:
          type: string
          example: Riquelme, Juan Román
        status:
          type: boolean
          example: true
        userType:
          type: array
          example: [Empleado, Titular de pago]
        roleUser:
          type: string
          example: user
        password:
          type: string
          example: El último 10
        email:
          type: string
          example: riquelme@bocajuniors.com
        phone:
          type: string
          example: 1123456789
        address:
          type: string
          example: Brandsen 805
...
    StandardResponse:
      type: object
      properties:
        status:
          type: number
          example: 200
        message:
          type: string
          example: 'Successful operation'
        data:
          anyOf:
            - type: object
            - type: array
      required:
        - status
        - message
        - data
...
    Users:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        cuil:
          type: string
        createdAt:
          type: string
          example: 2024-12-15T00:00:00.000Z
        updatedAt:
          type: string
          example: 2024-12-15T00:00:00.000Z
...
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
        status:
          type: number
          example: 500
        message:
          type: string
          example: Internal server error
...
    ErrorItem:
      type: object
      properties:
        key:
          type: string
          example: error.message.key
        message:
          type: string
          example: Error message example. Same as the outer 'message' property.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...